ported Buildcraft integration to the new compat system
