+++ src/main/java/buildcraftAdditions/compat/ModuleManager.java
@@ -9,6 +9,7 @@ import cpw.mods.fml.common.event.FMLLoadCompleteEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 
+import buildcraftAdditions.compat.buildcraft.CompatBC;
 import buildcraftAdditions.core.Logger;
 
 import com.google.common.base.Strings;
@@ -30,7 +31,8 @@ public class ModuleManager {
 	}
 
 	public void setupModules() {
-		registerModule(new CompatTest());
+		//registerModule(new CompatTest()); //debug compat.
+		registerModule(new CompatBC());
 	}
 
 	public void registerModule(Object object) {
