+++ src/main/java/buildcraftAdditions/compat/buildcraft/BCItems.java
@@ -1,4 +1,4 @@
-package buildcraftAdditions.utils;
+package buildcraftAdditions.compat.buildcraft;
 
 import net.minecraft.block.Block;
 import net.minecraft.item.Item;
