+++ src/main/java/buildcraftAdditions/compat/buildcraft/schematics/SchematicRotatableBCABlock.java
@@ -1,4 +1,4 @@
-package buildcraftAdditions.ModIntegration.Buildcraft.schematics;
+package buildcraftAdditions.compat.buildcraft.schematics;
 
 import java.util.LinkedList;
