+++ src/main/java/buildcraftAdditions/compat/buildcraft/schematics/SchematicSorter.java
@@ -1,11 +1,13 @@
-package buildcraftAdditions.ModIntegration.Buildcraft.schematics;
+package buildcraftAdditions.compat.buildcraft.schematics;
 
 import java.util.LinkedList;
 
 import net.minecraft.item.ItemStack;
 
 import buildcraft.api.blueprints.IBuilderContext;
-import buildcraft.api.blueprints.SchematicTile;
+import buildcraft.builders.schematics.SchematicPiston;
+
+import buildcraftAdditions.reference.ItemsAndBlocks;
 
 /**
  * Copyright (c) 2014-2015, AEnterprise
@@ -14,30 +16,17 @@ import buildcraft.api.blueprints.SchematicTile;
  * Please check the contents of the license located in
  * http://buildcraftadditions.wordpress.com/wiki/licensing-stuff/
  */
-public class SchematicBCABase extends SchematicTile {
+public class SchematicSorter extends SchematicPiston {
 
 	@Override
 	public void getRequirementsForPlacement(IBuilderContext context, LinkedList<ItemStack> requirements) {
-		requirements.add(new ItemStack(block, 1, 0));
-	}
-
-	@Override
-	public void storeRequirements(IBuilderContext context, int x, int y, int z) {
-		//don't store items that are in the machine
+		requirements.add(new ItemStack(ItemsAndBlocks.itemSorter));
 	}
 
 	@Override
 	public void placeInWorld(IBuilderContext context, int x, int y, int z, LinkedList<ItemStack> stacks) {
-		context.world().setBlock(x, y, z, block, meta, 2);
-	}
-
-	@Override
-	public void initializeFromObjectAt(IBuilderContext context, int x, int y, int z) {
-		//don't save TE data
-	}
-
-	@Override
-	public BuildingStage getBuildStage() {
-		return BuildingStage.STANDALONE;
+		super.placeInWorld(context, x, y, z, stacks);
+		tileNBT.setBoolean("reloadRotation", true);
+		context.world().getTileEntity(x, y, z).readFromNBT(tileNBT);
 	}
 }
