+++ src/main/java/buildcraftAdditions/compat/buildcraft/triggers/TriggerHasEmptyCanister.java
@@ -1,4 +1,4 @@
-package buildcraftAdditions.ModIntegration.Buildcraft.Triggers;
+package buildcraftAdditions.compat.buildcraft.triggers;
 
 import net.minecraft.tileentity.TileEntity;
 
@@ -7,7 +7,7 @@ import net.minecraftforge.common.util.ForgeDirection;
 import buildcraft.api.statements.IStatementContainer;
 import buildcraft.api.statements.IStatementParameter;
 
-import buildcraftAdditions.tileEntities.TileChargingStation;
+import buildcraftAdditions.tileEntities.TileFluidicCompressor;
 
 /**
  * Copyright (c) 2014, AEnterprise
@@ -16,17 +16,17 @@ import buildcraftAdditions.tileEntities.TileChargingStation;
  * Please check the contents of the license located in
  * http://buildcraftadditions.wordpress.com/wiki/licensing-stuff/
  */
-public class TriggerReadyToCharge extends BasicTrigger {
+public class TriggerHasEmptyCanister extends BasicTrigger {
 
-	public TriggerReadyToCharge() {
-		super("readyToCharge", "TriggerReadyToCharge");
+	public TriggerHasEmptyCanister() {
+		super("hasEmptyCanister", "TriggerHasEmptyCanister");
 	}
 
 	@Override
 	public boolean isTriggerActive(TileEntity target, ForgeDirection side, IStatementContainer source, IStatementParameter[] parameters) {
-		if (target instanceof TileChargingStation) {
-			TileChargingStation chargingStation = (TileChargingStation) target;
-			return chargingStation.getStackInSlot(0) == null;
+		if (target instanceof TileFluidicCompressor) {
+			TileFluidicCompressor fluidicCompressor = (TileFluidicCompressor) target;
+			return (fluidicCompressor.getStackInSlot(1) != null && fluidicCompressor.getStackInSlot(1).stackTagCompound != null && fluidicCompressor.getStackInSlot(1).stackTagCompound.hasKey("Fluid"));
 		}
 		return false;
 	}
