+++ src/main/java/buildcraftAdditions/compat/buildcraft/triggers/TriggerKEBEmpty.java
@@ -1,4 +1,4 @@
-package buildcraftAdditions.ModIntegration.Buildcraft.Triggers;
+package buildcraftAdditions.compat.buildcraft.triggers;
 
 import net.minecraft.tileentity.TileEntity;
 
@@ -10,20 +10,20 @@ import buildcraft.api.statements.IStatementParameter;
 import buildcraftAdditions.tileEntities.Bases.TileKineticEnergyBufferBase;
 
 /**
- * Copyright (c) 2014, AEnterprise
+ * Copyright (c) 2014-2015, AEnterprise
  * http://buildcraftadditions.wordpress.com/
- * Buildcraft Additions is distributed under the terms of GNU GPL v3.0
+ * Eureka is distributed under the terms of GNU GPL v3.0
  * Please check the contents of the license located in
  * http://buildcraftadditions.wordpress.com/wiki/licensing-stuff/
  */
-public class TriggerKEBCharged extends BasicTrigger {
+public class TriggerKEBEmpty extends BasicTrigger {
 
-	public TriggerKEBCharged() {
-		super("KEBCharged", "fullyCharged");
+	public TriggerKEBEmpty() {
+		super("KEBEmpty", "empty");
 	}
 
 	@Override
 	public boolean isTriggerActive(TileEntity target, ForgeDirection side, IStatementContainer source, IStatementParameter[] parameters) {
-		return target instanceof TileKineticEnergyBufferBase && ((TileKineticEnergyBufferBase) target).getEnergyLevel() == 100;
+		return target instanceof TileKineticEnergyBufferBase && ((TileKineticEnergyBufferBase) target).getEnergyLevel() == 0;
 	}
 }
