+++ src/main/java/buildcraftAdditions/compat/buildcraft/triggers/TriggerKEBUnder100.java
@@ -1,4 +1,4 @@
-package buildcraftAdditions.ModIntegration.Buildcraft.Triggers;
+package buildcraftAdditions.compat.buildcraft.triggers;
 
 import net.minecraft.tileentity.TileEntity;
 
@@ -7,7 +7,7 @@ import net.minecraftforge.common.util.ForgeDirection;
 import buildcraft.api.statements.IStatementContainer;
 import buildcraft.api.statements.IStatementParameter;
 
-import buildcraftAdditions.tileEntities.TileFluidicCompressor;
+import buildcraftAdditions.tileEntities.Bases.TileKineticEnergyBufferBase;
 
 /**
  * Copyright (c) 2014, AEnterprise
@@ -16,18 +16,14 @@ import buildcraftAdditions.tileEntities.TileFluidicCompressor;
  * Please check the contents of the license located in
  * http://buildcraftadditions.wordpress.com/wiki/licensing-stuff/
  */
-public class TriggerHasFullCanister extends BasicTrigger {
+public class TriggerKEBUnder100 extends BasicTrigger {
 
-	public TriggerHasFullCanister() {
-		super("hasFullCanister", "TriggerHasFullCanister");
+	public TriggerKEBUnder100() {
+		super("KEBUnder100", "lessThan100%");
 	}
 
 	@Override
 	public boolean isTriggerActive(TileEntity target, ForgeDirection side, IStatementContainer source, IStatementParameter[] parameters) {
-		if (target instanceof TileFluidicCompressor) {
-			TileFluidicCompressor fluidicCompressor = (TileFluidicCompressor) target;
-			return (fluidicCompressor.getStackInSlot(1) != null && fluidicCompressor.getStackInSlot(1).stackTagCompound.hasKey("Fluid"));
-		}
-		return false;
+		return target instanceof TileKineticEnergyBufferBase && ((TileKineticEnergyBufferBase) target).getEnergyLevel() < 100;
 	}
 }
