+++ src/main/java/buildcraftAdditions/compat/buildcraft/triggers/TriggerKEBUnder50.java
@@ -1,4 +1,4 @@
-package buildcraftAdditions.ModIntegration.Buildcraft.Triggers;
+package buildcraftAdditions.compat.buildcraft.triggers;
 
 import net.minecraft.tileentity.TileEntity;
 
@@ -10,20 +10,20 @@ import buildcraft.api.statements.IStatementParameter;
 import buildcraftAdditions.tileEntities.Bases.TileKineticEnergyBufferBase;
 
 /**
- * Copyright (c) 2014-2015, AEnterprise
+ * Copyright (c) 2014, AEnterprise
  * http://buildcraftadditions.wordpress.com/
- * Eureka is distributed under the terms of GNU GPL v3.0
+ * Buildcraft Additions is distributed under the terms of GNU GPL v3.0
  * Please check the contents of the license located in
  * http://buildcraftadditions.wordpress.com/wiki/licensing-stuff/
  */
-public class TriggerKEBEmpty extends BasicTrigger {
+public class TriggerKEBUnder50 extends BasicTrigger {
 
-	public TriggerKEBEmpty() {
-		super("KEBEmpty", "empty");
+	public TriggerKEBUnder50() {
+		super("KEBUnder50", "lessThan50%");
 	}
 
 	@Override
 	public boolean isTriggerActive(TileEntity target, ForgeDirection side, IStatementContainer source, IStatementParameter[] parameters) {
-		return target instanceof TileKineticEnergyBufferBase && ((TileKineticEnergyBufferBase) target).getEnergyLevel() == 0;
+		return target instanceof TileKineticEnergyBufferBase && ((TileKineticEnergyBufferBase) target).getEnergyLevel() < 50;
 	}
 }
