+++ src/main/java/buildcraftAdditions/compat/buildcraft/triggers/TriggerKEBUnder75.java
@@ -1,4 +1,4 @@
-package buildcraftAdditions.ModIntegration.Buildcraft.Triggers;
+package buildcraftAdditions.compat.buildcraft.triggers;
 
 import net.minecraft.tileentity.TileEntity;
 
@@ -16,14 +16,14 @@ import buildcraftAdditions.tileEntities.Bases.TileKineticEnergyBufferBase;
  * Please check the contents of the license located in
  * http://buildcraftadditions.wordpress.com/wiki/licensing-stuff/
  */
-public class TriggerKEBUnder50 extends BasicTrigger {
+public class TriggerKEBUnder75 extends BasicTrigger {
 
-	public TriggerKEBUnder50() {
-		super("KEBUnder50", "lessThan50%");
+	public TriggerKEBUnder75() {
+		super("KEBUnder75", "lessThan75%");
 	}
 
 	@Override
 	public boolean isTriggerActive(TileEntity target, ForgeDirection side, IStatementContainer source, IStatementParameter[] parameters) {
-		return target instanceof TileKineticEnergyBufferBase && ((TileKineticEnergyBufferBase) target).getEnergyLevel() < 50;
+		return target instanceof TileKineticEnergyBufferBase && ((TileKineticEnergyBufferBase) target).getEnergyLevel() < 75;
 	}
 }
