+++ src/main/java/buildcraftAdditions/compat/buildcraft/triggers/TriggerProvider.java
@@ -1,4 +1,4 @@
-package buildcraftAdditions.ModIntegration.Buildcraft.Triggers;
+package buildcraftAdditions.compat.buildcraft.triggers;
 
 import java.util.ArrayList;
 import java.util.Collection;
